feat: instruction for booting and application

+++ src/main/java/com/example/libraryservice/util/ConversionUtil.java
@@ -3,7 +3,7 @@ package com.example.libraryservice.util;
 public class ConversionUtil {
     public static Integer convertStringToInteger(String str) {
         try {
-            return (Integer) (int) Double.parseDouble(str);
+            return (int) Double.parseDouble(str);
         } catch (NumberFormatException e) {
             return null;
         }
